+++ ChangeLog.md
@@ -8,6 +8,10 @@ Blob:
 
 - Fix validation of Blob SAS token when using the second key for an account in `AZURITE_ACCOUNTS`
 
+Table:
+
+- Fixed the errorCode returned, when malformed Etag is provided for table Update/Delete calls. (issue #2013)
+
 ## 2023.08 Version 3.26.0
 
 General: